Add fallback batch constructor from scalars + clean up
- Allow fallback batches to be created from a list of scalars
- Merge implementation details into xsimd_utils.hpp
- Use default constructor instead of 0 in broadcasting array constructor
- Turn to_unsigned and from_unsigned into functions
